feat(I132): remove replace `/home` by `/ `
