--- src/app/home/page.view.tsx
@@ -1,12 +1,0 @@
-export default function HomePage() {
-  return (
-    <div className="min-h-[70vh] flex flex-col items-center justify-center animate-fade-in-up">
-      <div className="text-center max-w-md">
-        <h1 className="text-4xl font-bold mb-4 tracking-tight">Home Page</h1>
-        <p className="text-gray-500 leading-relaxed text-lg">
-          Esta página está temporariamente vazia. Em breve novidades por aqui!
-        </p>
-      </div>
-    </div>
-  );
-}
