+++ src/app/page.view.tsx
@@ -1,14 +1,12 @@
-import Link from "next/link";
-
-export default function Home() {
+export default function HomePage() {
   return (
-    <div className="font-sans grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
-      <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
-        <Link href="/timeline">Go to User Page</Link>
-      </main>
-      <footer className="row-start-3 flex gap-[24px] flex-wrap items-center justify-center">
-        footer
-      </footer>
+    <div className="min-h-[70vh] flex flex-col items-center justify-center animate-fade-in-up">
+      <div className="text-center max-w-md">
+        <h1 className="text-4xl font-bold mb-4 tracking-tight">Home Page</h1>
+        <p className="text-gray-500 leading-relaxed text-lg">
+          Esta página está temporariamente vazia. Em breve novidades por aqui!
+        </p>
+      </div>
     </div>
   );
 }
